проект

+++ src/components/card.js
@@ -1,6 +1,7 @@
 export {createCard, deleteCard }
+import { likeCardServer, unLikeCardServer, deleteCardServer } from './api.js'
 
-function createCard(linkValue, nameValue, deleteCard, openCardImage ) {
+function createCard(linkValue, nameValue, deleteCard, openCardImage, userId, ownerId, cardId, likes, myLikes) {
     const cardTemplate = document.querySelector('#card-template').content;
     const cardElement = cardTemplate.querySelector('.card').cloneNode(true);
     const cardDeleteButton = cardElement.querySelector('.card__delete-button');
@@ -9,28 +10,54 @@ function createCard(linkValue, nameValue, deleteCard, openCardImage ) {
     cardImg.src = linkValue;
     cardImg.alt = nameValue;
     cardTitle.textContent = nameValue;
+    if (userId === ownerId) {
+        cardDeleteButton.style.display = 'inline-block';
+        cardDeleteButton.addEventListener('click', () => deleteCard(cardElement));
+        cardDeleteButton.addEventListener('click', () => deleteCardServer(cardId));
+    }
+    else {
+        cardDeleteButton.style.display = 'none'
+    }
 
-
-    cardDeleteButton.addEventListener('click', () => deleteCard(cardElement));
     // лайк карточки
     const likeButton = cardElement.querySelector('.card__like-button');
+    if (myLikes) likeButton.classList.toggle('card__like-button_is-active')
+    let likeCount = cardElement.querySelector('.like-count');
+    likeCount.textContent = likes ? likes.length : 0;
+    likeButton.addEventListener('click', (evt) => handleLikeClick(evt, cardId, likeCount));
 
-    likeButton.addEventListener('click', likeCard);
+    // открытие карточки
+cardImg.addEventListener('click', () => {
+    openCardImage(cardImg.src, cardTitle.textContent)
+});
 
-    cardImg.addEventListener('click', () => {
-        openCardImage(cardImg.src, cardTitle.textContent)
-    });
-
-    return cardElement;
+return cardElement;
 };
 
 // функция удаление карточки
-
 function deleteCard (el) {
     el.remove();
 };
-// функция лайка карточки
 
-function likeCard (evt) {
-    evt.target.classList.toggle('card__like-button_is-active')
-};
+// функция лайка карточки
+function handleLikeClick(evt, cardId, likeCount) {
+    if(!evt.target.classList.contains('card__like-button_is-active')) {
+        likeCardServer(cardId)
+        .then((result) => {
+            likeCount.textContent = result.likes.length;
+            evt.target.classList.toggle('card__like-button_is-active');
+        })
+        .catch((err) => {
+            console.log(err);
+        })
+    } else {
+        unLikeCardServer(cardId)
+        .then((result) => {
+            likeCount.textContent = result.likes.length;
+            evt.target.classList.toggle('card__like-button_is-active');
+        })
+        .catch((err) => {
+            console.log(err);
+        })
+    }
+}
